fixing the url that data gets posted to
we now use a full url to post data instead of a relative url path because running this application in a container with a relative url path did not seem to send post the request back to the correct web server. the js log in the browser showed the post request was made, but the wrwroxy proxy never showed the request making it that far. trying to make the request using a full url in curl did work.